fix: Fix incorrect reflections
The first part of the issue was raycasts returning incorrect normals -
after applying the inverse transform (object to world space), the result
wasn't normalized.
The second part of the issue was a wrong equation for vector reflection,
since it seemed to be working well with the previously mentioned
incorrect normals.